Rename FastaSaver -> CSVSaver and write sequences
Write per sequence results to a file as well as the per loop result. The loop
result contains the aggregated results for the loop while per sequence contains
the information for each loop in the sequence. The classes is also renamed
appropriately. We also now write headers and format the numbers in the CSV file
for ease of use.