Remove numSequences from core JAR3D.
The core JAR3D uses numSequences to silently truncate the number of sequences
analyzed. It really didn't so much truncate as just not process all inputs.
Since the sData variable is both input and output it's possible to see odd
things where not all inputs have a result because of a bad setting in
numSequences.  I've removed that. It will now use all sequences given. In
JAR3DMatlab however it still uses numSequences to silently truncate. I don't
want to break any behavior. I will check if this is needed.