Major changes.
This is a long overdue commit that introduces many changes to JAR3D. Not least
of which is building this with Maven. The project can now be built and tested
with Maven. In addition I've made many changes to how results are created. This
is to allow for more useful saving of results.

I have also added lots of documentation and removed a lot of trailing whitespace.