Clean up and add profiles to pom.
The clean up was just some indenting and spacing issues to make it more
readable.

I've also added a profiles section to the pom. This means we can build two
different jars in different enviroments. By deafult it builds a jar which
creates a csv. When given the -P db option it will create a jar that accesses
the DB.