Correctly package with dependencies.
I was not using the assembly plugin correctly before. I've fixed it and now it
correctly generates a jar with everything in it. I've also made this part of
the normal build process for packaging, so it should be possible to create
this jar by doing mvn clean package -P db.