Hairpin node correspondences
When the sequence is so short that there are not enough letters to fill
all of the fixed positions in a hairpin, the max log probability score
will be horrible, but hairpins cannot be deleted and ShowCorrespondences
needs to say what each position in the sequence corresponds to.  So I
made it so that each sequence position that is assigned to the hairpin
corresponds to position 1.  That will stand out as looking wrong, but I
can't see a perfect solution.